Refactored getCalendarSuiteGetFile:
create a new struct for the calendarSuite called calendarSuite, holds an array of pointer to calendars and the length of the array. Implemented this in the rest of the code.

printMetadataCalendarSuite new only needs one argument: an calendarSuite.

Created a function initCalendarSuite, hvis creates a calendarSuite sized to fit the input of *.ics files. idk but this might be causing some kind of memory issue :^)